Partial unpack fix

+++ unpack.sh
@@ -19,11 +19,17 @@ function lfs_pull() {
 }
 
 function unpack() {
+    echo "Unpacking data/ ..."
     find ./data -type f -iname "*.xz" -exec xz -d -k {} \;
 }
 
 function filter() {
     local fname=$1
+    if [ ! -f ${fname} ]; then
+        echo "Cannot filter non-existent file ..."
+        exit 3
+    fi
+    echo "Filtering ${fname} ..."
     cat ${fname}|egrep -v '^[\.|-|%]' > ${fname}.1
     mv ${fname}.1 ${fname}
 }
@@ -32,7 +38,9 @@ function combine() {
     olddir=$(pwd)
     for datadir in $(find ./data -type d -mindepth 1); do
         cd ${datadir}
+        echo "Working on ${datadir} ..."
         big_fname=$(ls *.txt|sed -E 's/[0-9]+\./\./g'|head -n 1)
+        rm -f ${big_fname}
         for fname in $(ls *.txt|grep '[0-9]\.txt'); do
             cat $fname >> $big_fname
             rm $fname
